Made maps actual hash-maps instead of linear arrays
This should let us have giant maps without performance penalty.
There are a couple of things to investigate still. The primary one
is how to handle rehashing, we should probably be doubling the
size of the hash map instead of adding a fixed amount, that way when we
get really large arrays, we don't have constant rehashing operations,
which are significantly more expensive than the dynamic array approach.